requirements.txt
absl-py==2.0.0
accelerate==0.25.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.1
aioredis==2.0.1
aiosignal==1.3.1
albumentations==1.1.0
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blind-watermark==0.4.4
blinker==1.7.0
blis==0.7.11
blobfile==2.1.1
boto3==1.34.106
botocore==1.34.106
braceexpand==0.1.7
brokenaxes==0.6.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
-e git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33#egg=clip
clip-interrogator==0.6.0
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
confection==0.1.4
contextlib2==21.6.0
contexttimer==0.3.3
controlnet-aux==0.0.7
crcmod==1.7
cryptography==42.0.5
ctgan==0.10.0
ctranslate2==4.1.0
cycler==0.12.1
cymem==2.0.8
datasets==2.18.0
debugpy==1.8.0
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
diffusers==0.25.1
dill==0.3.8
distlib==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
dominate==2.9.1
einops==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
facexlib==0.3.0
fairscale==0.4.4
Faker==24.9.0
fastapi==0.108.0
fastjsonschema==2.19.1
ffmpy==0.3.1
filelock==3.12.2
filterpy==1.4.5
fire==0.6.0
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.47.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.5.4
gdown==5.1.0
gfpgan==1.3.8
gitdb==4.0.11
GitPython==3.1.40
glog==0.3.1
google==3.0.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
gradio==4.13.0
gradio_client==0.8.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.23.2
humanfriendly==10.0
icecream==2.1.3
identify==2.5.33
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
imgaug==0.4.0
importlib-metadata==6.11.0
importlib-resources==6.1.1
imutils==0.5.4
iopath==0.1.10
ipykernel==6.28.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json-tricks==3.17.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kaggle==1.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kmeans-pytorch==0.3
kornia==0.7.1
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.10.0
littleutils==0.2.2
lmdb==1.4.1
loguru==0.7.2
loralib==0.1.2
lpips==0.1.4
lxml==4.9.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mediapipe==0.10.11
mistune==3.0.2
ml-dtypes==0.2.0
modelscope==1.13.1
mpi4py==3.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
munch==4.0.0
murmurhash==1.0.10
natsort==8.4.0
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
ninja==1.11.1.1
nltk==3.8.1
nni==3.0
nodeenv==1.8.0
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.0
nvidia-ml-py==12.535.133
oauthlib==3.2.2
ogb==1.3.6
omegaconf==2.3.0
onnxruntime==1.17.3
open-clip-torch==2.24.0
openai==1.6.1
openai-clip==1.0.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.10
oss2==2.18.4
outdated==0.2.2
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pbr==6.0.0
pillow==10.2.0
platformdirs==4.1.0
plotly==5.18.0
portalocker==2.8.2
preshed==3.0.9
pretrainedmodels==0.7.4
prettytable==3.9.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
pudb==2023.1
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyiqa==0.1.11
PyMySQL==1.1.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gflags==3.1.2
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-slugify==8.0.1
PythonWebHDFS==0.2.3
pytorch-gan-metrics==0.5.3
pytorch-lightning==2.2.1
pytorch-pretrained-biggan==0.1.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qiskit==1.2.0
qiskit-aer==0.15.0
QtAwesome==1.3.0
qtconsole==5.5.1
QtPy==2.4.1
qudida==0.0.4
rdt==1.11.0
realesrgan==0.3.0
referencing==0.32.1
regex==2023.12.25
repackage==0.7.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.24.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
rustworkx==0.15.1
s3transfer==0.10.1
sacremoses==0.1.1
safetensors==0.4.1
schema==0.7.5
scikit-image==0.18.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.11.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.45.0
setproctitle==1.3.3
shapely==2.0.5
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sk-video==1.1.10
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.32.0.post1
statannot==0.2.3
statannotations==0.6.0
stevedore==5.3.0
streamlit==1.29.0
supervision==0.19.0
symengine==0.11.0
sympy==1.12
tabulate==0.9.0
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.0
test-tube==0.7.5
text-unidecode==1.3
tgan==0.0.0
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.2.1+cu118
torch-fidelity==0.3.0
torch-kmeans==0.2.0
torchaudio==2.2.1+cu118
torchcam==0.4.0
torchfile==0.1.0
torchmetrics==1.2.1
torchsummary==1.5.1
torchvision==0.17.1+cu118
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.2
typeguard==4.1.2
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.1
urwid==2.4.1
urwid-readline==0.13
uvicorn==0.25.0
validators==0.22.0
virtualenv==20.25.0
wandb==0.16.6
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webdataset==0.2.86
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wilds==2.0.0
win32-setctime==1.1.0
wrapt==1.14.1
xgboost==2.0.3
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
